update copyrights headr and docs
