--- src/foam/lib/parse/Eof.java
@@ -1,37 +1,0 @@
-/**
- * @license
- * Copyright 2016 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package foam.lib.parse;
-
-import foam.lib.json.Whitespace;
-
-public class Eof
-  implements Parser
-{
-  protected static final Parser instance__ = new Eof();
-
-  public static Parser instance() {
-    return instance__;
-  }
-
-  private Eof() {
-  }
-
-  public PStream parse(PStream ps, ParserContext x) {
-    return ps.valid() ? null : ps;
-  }
-}
