+++ src/foam/lib/parse/GreedyAlt.java
@@ -1,23 +1,16 @@
 /**
- * @license
- * Copyright 2016 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+
+ @license
+ Copyright 2021 The FOAM Authors. All Rights Reserved.
+ http://www.apache.org/licenses/LICENSE-2.0
  */
 
 package foam.lib.parse;
 
-
+/**
+Applies all parsers and returns PStream produced by the parser that consumes most of the characters
+ opposed to Alt parser that returns first matched value.
+ */
 public class GreedyAlt
   implements Parser
 {
